requirements.txt
requests
Pillow
google-cloud-storage
python-dotenv
apify-client
